Moved the env to its own file.
