+++ frispy/__init__.py
@@ -1,4 +1,5 @@
-from .frispy import Disc, Environment
+from environment import Environment
+from frispy import Disc
 
 __author__ = "Tom McClintock thmsmcclintock@gmail.com"
 __version__ = "1.0.0"
